wip release (#3)
* fix: add base handlers for amqp, add health check

* fix: remove http handler for health check (use middleware)

* fix: health check

* fix: serve amqp health-check

* fix: amqp connection timeout

* fix

* fix: zaebal krol

* fix: ...

* fix: amqp errors

* fix: bugs, init fcm and apn adapters, add notification domain

* fix: add handlers, fix start bugs

* fix: impl notifier

* fix: parse port for SMTP config

* fix: minor refactor adapters (fcn, apn), add template support for email and push

* fix: add mongo support

* fix: go mod tidy

+++ internal/config/smtp.go
@@ -9,7 +9,7 @@ import (
 type SMTPConfig struct {
 	From     string `env:"SMTP_FROM_EMAIL"`
 	Host     string `env:"SMTP_HOST"`
-	Port     int8   `env:"SMTP_PORT"`
+	Port     int16  `env:"SMTP_PORT"`
 	Username string `env:"SMS_USERNAME"`
 	Password string `env:"SMS_PASSWORD"`
 }
